+++ src/lib/pages/validator-details/components/validator-top/WebsiteButton.tsx
@@ -1,17 +1,23 @@
-import { Button } from "@chakra-ui/react";
+import { Button, Link } from "@chakra-ui/react";
 import type { ButtonProps } from "@chakra-ui/react";
 
 import { CustomIcon } from "lib/components/icon";
 
-export const WebsiteButton = (props: ButtonProps) => (
-  <Button
-    leftIcon={<CustomIcon name="website" boxSize={3} />}
-    rightIcon={<CustomIcon name="launch" boxSize={3} />}
-    size="sm"
-    variant="outline-primary"
-    width={{ base: "full", md: "auto" }}
-    {...props}
-  >
-    Website
-  </Button>
+interface WebsiteButtonProps extends ButtonProps {
+  href: string;
+}
+
+export const WebsiteButton = ({ href, ...props }: WebsiteButtonProps) => (
+  <Link href={href} isExternal _hover={{ textDecoration: "none" }}>
+    <Button
+      leftIcon={<CustomIcon name="website" boxSize={3} />}
+      rightIcon={<CustomIcon name="launch" boxSize={3} />}
+      size="sm"
+      variant="outline-primary"
+      width={{ base: "full", md: "auto" }}
+      {...props}
+    >
+      Website
+    </Button>
+  </Link>
 );