+++ src/lib/services/delegationService.ts
@@ -3,7 +3,7 @@ import { useQuery } from "@tanstack/react-query";
 import {
   CELATONE_QUERY_KEYS,
   useBaseApiRoute,
-  useCelatoneApp,
+  useGovConfig,
 } from "lib/app-provider";
 import type { BechAddr } from "lib/types";
 
@@ -12,17 +12,13 @@ import { getDelegationsByAddress } from "./delegation";
 
 export const useDelegationsByAddress = (address: BechAddr) => {
   const endpoint = useBaseApiRoute("accounts");
-  const {
-    chainConfig: {
-      extra: { disableDelegation },
-    },
-  } = useCelatoneApp();
+  const { enabled } = useGovConfig({ shouldRedirect: false });
 
   return useQuery(
     [CELATONE_QUERY_KEYS.DELEGATIONS_BY_ADDRESS, endpoint, address],
     () => getDelegationsByAddress(endpoint, address),
     {
-      enabled: !disableDelegation,
+      enabled,
       refetchOnWindowFocus: false,
     }
   );