+++ src/lib/services/validator.ts
@@ -6,13 +6,18 @@ import { CURR_THEME } from "env";
 import type { Option, StakingShare, Validator, ValidatorAddr } from "lib/types";
 import {
   BlockVote,
+  zBechAddr,
   zBig,
   zCoin,
   zUtcDate,
-  zValidatorAddr,
   zValidatorData,
 } from "lib/types";
-import { parseWithError, removeSpecialChars, snakeToCamel } from "lib/utils";
+import {
+  parseTxHash,
+  parseWithError,
+  removeSpecialChars,
+  snakeToCamel,
+} from "lib/utils";
 
 import { zBlocksResponse } from "./block";
 import { zProposal } from "./proposal";
@@ -186,12 +191,12 @@ export const getValidatorData = async (
 const zValidatorUptimeResponse = z
   .object({
     uptime: z.object({
-      signed_block: z.number(),
-      proposed_block: z.number(),
+      signed_blocks: z.number(),
+      proposed_blocks: z.number(),
       missed_blocks: z.number(),
       total: z.number(),
     }),
-    recent_blocks: z
+    recent_100_blocks: z
       .object({
         height: z.number(),
         vote: z.nativeEnum(BlockVote),
@@ -223,13 +228,21 @@ export const getValidatorUptime = async (
 
 const zValidatorDelegationRelatedTxsResponseItem = z
   .object({
-    tx_hash: z.string(),
+    tx_hash: z.string().transform(parseTxHash),
     height: z.number().positive(),
     tokens: zCoin.array(),
     timestamp: zUtcDate,
-    validator_address: zValidatorAddr,
+    messages: z.array(
+      z.object({
+        type: z.string(),
+      })
+    ),
+    sender: zBechAddr,
   })
   .transform(snakeToCamel);
+export type ValidatorDelegationRelatedTxsResponseItem = z.infer<
+  typeof zValidatorDelegationRelatedTxsResponseItem
+>;
 
 const zValidatorDelegationRelatedTxsResponse = z.object({
   items: z.array(zValidatorDelegationRelatedTxsResponseItem),