+++ src/lib/services/validatorService.ts
@@ -158,10 +158,11 @@ export const useValidatorUptime = (
   const endpoint = useBaseApiRoute("validators");
 
   return useQuery<ValidatorUptimeResponse>(
-    [CELATONE_QUERY_KEYS.VALIDATOR_UPTIME, endpoint, validatorAddress],
+    [CELATONE_QUERY_KEYS.VALIDATOR_UPTIME, endpoint, validatorAddress, blocks],
     async () => getValidatorUptime(endpoint, validatorAddress, blocks),
     {
       retry: 1,
+      refetchOnWindowFocus: false,
     }
   );
 };
@@ -169,7 +170,11 @@ export const useValidatorUptime = (
 export const useValidatorDelegationRelatedTxs = (
   validatorAddress: ValidatorAddr,
   limit: number,
-  offset: number
+  offset: number,
+  options: Pick<
+    UseQueryOptions<ValidatorDelegationRelatedTxsResponse>,
+    "onSuccess"
+  > = {}
 ) => {
   const endpoint = useBaseApiRoute("validators");
 
@@ -178,6 +183,8 @@ export const useValidatorDelegationRelatedTxs = (
       CELATONE_QUERY_KEYS.VALIDATOR_DELEGATION_RELATED_TXS,
       endpoint,
       validatorAddress,
+      limit,
+      offset,
     ],
     async () =>
       getValidatorDelegationRelatedTxs(
@@ -186,7 +193,7 @@ export const useValidatorDelegationRelatedTxs = (
         limit,
         offset
       ),
-    { retry: 1 }
+    { retry: 1, ...options }
   );
 };